refactor: move GetBEEF logic to separate file
